refactor(lazer): add index.js and reexport solana contract idl
